+++ src/index.tsx
@@ -1,6 +1,13 @@
 import CSSMotion, { CSSMotionProps } from './CSSMotion';
 import CSSMotionList, { CSSMotionListProps } from './CSSMotionList';
+import { MotionEventHandler, MotionEndEventHandler } from './interface';
 
-export { CSSMotionProps, CSSMotionList, CSSMotionListProps };
+export {
+  CSSMotionProps,
+  CSSMotionList,
+  CSSMotionListProps,
+  MotionEventHandler,
+  MotionEndEventHandler,
+};
 
 export default CSSMotion;
